refactored job, intro, about, and contact styling rel: #100
